feat: bake the glove's texture and share the silhouette between both stages
Silhouette measurement and the descriptor move to forge/_shared/glove_silhouette.py.
Stage 2 decides what geometry the spec declares and stage 3 builds it, so both need
the same measurement and the same descriptor shape; keeping it shared means neither
stage imports the other and the two cannot disagree about which view was measured.

bake_shell_atlas turns the admitted plates into the texture uv0 already addresses:
each plate cropped to its measured component bounding box, the same box the grid was
resampled from. Cropping the plate half instead leaves background wherever the glove
does not fill it, which reads as a black halo on the model. The bundle carries the
atlas record and the descriptor inside rootDigest, so the browser cannot be handed a
texture or a geometry the review did not see.

The atlas also bleeds the glove's colour twelve pixels past the silhouette before
cropping. Without that, a texel sampled at the boundary averages in the plate's
background. This was only visible in a real browser at above-capture resolution --
the 1024 Playwright capture hid it, and driving the page over CDP is what surfaced
both it and a 404 that Playwright reported as no console error at all.

Corrects a claim from ca94beb. That commit said the e2e harness now runs in the
suite. It does here and not on a fresh clone: .gitignore excludes runtime/* and
re-includes only runtime/scripts/, so runtime/glove-review is absent from the
repository even though it is source rather than capture output. The harness test now
skips with that reason named, instead of failing with a bare file-not-found that
reads as a broken harness. Whether that ignore rule should be narrowed is the
owner's call; the rule's own comment describes it as excluding capture output.

802 tests pass with the showcase gates skipped; the failing set with them enabled
still contains nothing absent from 7dca2c4.
