+++ forge/tests/test_glove_e2e_harness.py
@@ -21,6 +21,17 @@ from forge.stage4_review.glove_review import METRIC_KINDS, REPORT_VERSION
 from forge.tests.run_glove_e2e import run_golden
 
 
+GLOVE_RUNTIME = ROOT / "runtime" / "glove-review" / "src" / "index.mjs"
+
+
+@unittest.skipUnless(
+    GLOVE_RUNTIME.is_file(),
+    # .gitignore excludes runtime/* and re-includes only runtime/scripts/, so runtime/glove-review/
+    # is absent from a fresh clone even though it is source rather than capture output. Without this
+    # guard the harness fails here with a bare file-not-found, which reads as a broken harness
+    # instead of a missing runtime.
+    f"{GLOVE_RUNTIME.relative_to(ROOT)} is absent: runtime/glove-review is not tracked by git",
+)
 class GloveE2EHarnessTests(unittest.TestCase):
     def test_documented_golden_invocation_runs_and_writes_a_report(self):
         with tempfile.TemporaryDirectory(prefix="img2-glove-harness-") as raw:
